Refactors backend and frontend code for better structure
Improves code organization and maintainability by:

- Introducing constants for agent IDs, dependency files, and language mappings in the backend
- Refactoring the AzureAgentService to utilize these constants and improve mock data generation
- Improving type hinting for better code clarity
- Restructuring the frontend with separate components for HomePage, RepoPage, and AgentCard
- Moving agent data into a separate file and creating utility functions for agent-related tasks

These changes result in a more modular and maintainable codebase for both the backend and frontend.
