+++ backend/requirements-dev.txt
@@ -1,0 +1,5 @@
+black==23.3.0
+isort==5.12.0
+flake8==6.0.0
+mypy==1.3.0
+pytest==7.3.1
