+++ frontend/src/vite-env.d.ts
@@ -1,0 +1,8 @@
+
+interface ImportMetaEnv {
+  readonly VITE_API_URL: string;
+}
+
+interface ImportMeta {
+  readonly env: ImportMetaEnv;
+}
